refactor(map): moved post map into components
