+++ lib/views/pages/home/content/map/map_screen.dart
@@ -4,7 +4,7 @@ import "package:proxima/viewmodels/map/map_view_model.dart";
 import "package:proxima/views/components/async/circular_value.dart";
 import "package:proxima/views/components/async/error_refresh_page.dart";
 import "package:proxima/views/components/options/map/map_selection_option_chips.dart";
-import "package:proxima/views/pages/home/content/map/post_map.dart";
+import "package:proxima/views/pages/home/content/map/components/post_map.dart";
 
 /// This widget displays a map with chips to select the type of map.
 class MapScreen extends ConsumerWidget {
