+++ test/end2end/app_actions.dart
@@ -7,7 +7,7 @@ import "package:proxima/views/pages/create_account/create_account_page.dart";
 import "package:proxima/views/pages/home/content/feed/components/post_card.dart";
 import "package:proxima/views/pages/home/content/feed/post_feed.dart";
 import "package:proxima/views/pages/home/content/map/map_screen.dart";
-import "package:proxima/views/pages/home/content/map/post_map.dart";
+import "package:proxima/views/pages/home/content/map/components/post_map.dart";
 import "package:proxima/views/pages/home/content/ranking/components/ranking_widget.dart";
 import "package:proxima/views/pages/home/home_page.dart";
 import "package:proxima/views/pages/home/home_top_bar/home_top_bar.dart";
